pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "be-auto-viral"
version = "0.1.0"
description = "AI Series Dashboard - Premium SaaS backend for faceless content automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    # Pin bcrypt for passlib compatibility (bcrypt 4.x breaks passlib on some setups)
    "bcrypt<4",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "stripe>=8.0.0",
    "httpx>=0.26.0",
    "python-multipart",
    "python-dateutil",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
